# schemas/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rideshare Simulation Control Panel API",
    "description": "REST API for controlling simulation and streaming real-time updates",
    "version": "1.0.0"
  },
  "paths": {
...
    "/simulation/status": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Get Status",
        "description": "Get current simulation status with detailed agent counts.",
        "operationId": "get_status_simulation_status_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/validate": {
      "get": {
        "summary": "Validate Api Key Endpoint",
        "description": "Validate API key for login.\n\nThis endpoint requires a valid API key and returns 200 if valid.\nUsed by the frontend login screen to validate credentials.\nReturns 401 if the API key is invalid.",
        "operationId": "validate_api_key_endpoint_auth_validate_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Validate Api Key Endpoint Auth Validate Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SimulationStatusResponse": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "stopped",
              "running",
              "draining",
              "paused"
            ],
            "title": "State"
          },
          "speed_multiplier": {
            "type": "integer",
            "title": "Speed Multiplier"
          },
          "current_time": {
            "type": "string",
            "title": "Current Time"
          },
          "drivers_total": {
            "type": "integer",
            "title": "Drivers Total"
          },
          "drivers_offline": {
            "type": "integer",
            "title": "Drivers Offline"
          },
          "drivers_online": {
            "type": "integer",
            "title": "Drivers Online"
          },
          "drivers_en_route_pickup": {
            "type": "integer",
            "title": "Drivers En Route Pickup"
          },
          "drivers_en_route_destination": {
            "type": "integer",
            "title": "Drivers En Route Destination"
          },
          "riders_total": {
            "type": "integer",
            "title": "Riders Total"
          },
          "riders_offline": {
            "type": "integer",
            "title": "Riders Offline"
          },
          "riders_waiting": {
            "type": "integer",
            "title": "Riders Waiting"
          },
          "riders_in_trip": {
            "type": "integer",
            "title": "Riders In Trip"
          },
          "active_trips_count": {
            "type": "integer",
            "title": "Active Trips Count"
          },
          "uptime_seconds": {
            "type": "number",
            "title": "Uptime Seconds"
          }
        },
        "type": "object",
        "required": [
          "state",
          "speed_multiplier",
          "current_time",
          "drivers_total",
          "drivers_offline",
          "drivers_online",
          "drivers_en_route_pickup",
          "drivers_en_route_destination",
          "riders_total",
          "riders_offline",
          "riders_waiting",
          "riders_in_trip",
          "active_trips_count",
          "uptime_seconds"
        ],
        "title": "SimulationStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}